scripts/requirements.txt
supabase==2.5.0
requests==2.31.0
python-dotenv==1.0.1
Pillow>=11.0.0
